refactor(libsy): share tier test fixtures and fold away TierClassifier
