+++ crates/libsy/src/lib.rs
@@ -15,7 +15,7 @@ pub use error::{DriverError, LibsyError, Result};
 
 mod algorithms;
 pub use algorithms::advisor_gate::{AdvisorGate, AdvisorGateConfig, GateTrigger};
-pub use algorithms::hierarchical::{HierarchicalRouter, HierarchicalRouterConfig, TierClassifier};
+pub use algorithms::hierarchical::{HierarchicalRouter, HierarchicalRouterConfig};
 pub use algorithms::llm_class::{
     CustomClassifierConfig, CustomClassifierPolicy, LlmClassifierConfig, LlmTaskClassifier,
     TaskClassifierConfig,
